Moved project files to top-level directory
