+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import { ChakraProvider } from '@chakra-ui/react'
 
-import theme from '../theme'
+import theme from '../themes/theme'
 import { AppProps } from 'next/app'
 
 function MyApp({ Component, pageProps }: AppProps) {
